Moved popup out of its own folder
